+++ src/cli/commands/analytics.py
@@ -3,7 +3,16 @@ from __future__ import annotations
 import argparse
 import asyncio
 
+import typer
+
 from src.cli import runtime
+from src.cli.commands.common import (
+    _NEG_ID_POSITIONAL,
+    AnalyticsGenre,
+    AnalyticsUseful,
+    apply_startup,
+    run_async,
+)
 
 # --------------------------------------------------------------------------- #
 # Per-action impls (epic #959, Wave 4 — issue #1124)
@@ -529,3 +538,183 @@ def run(args: argparse.Namespace) -> None:
         )
 
     asyncio.run(coro)
+
+
+# --------------------------------------------------------------------------- #
+# analytics → top / content-types / hourly / summary / daily / pipeline-stats /
+#   trending-topics / trending-channels / velocity / peak-hours / calendar /
+#   trending-emojis / channel / channel-rating / channel-rate
+# --------------------------------------------------------------------------- #
+
+analytics_app = typer.Typer(no_args_is_help=True, help="Message analytics")
+
+
+@analytics_app.command("top")
+def analytics_top(
+    ctx: typer.Context,
+    limit: int = typer.Option(20, "--limit", help="Number of results (default: 20)"),
+    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
+    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
+) -> None:
+    """Top messages by reactions."""
+    apply_startup(ctx)
+    run_async(top_impl(ctx.obj.config, limit=limit, date_from=date_from, date_to=date_to))
+
+
+@analytics_app.command("content-types")
+def analytics_content_types(
+    ctx: typer.Context,
+    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
+    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
+) -> None:
+    """Engagement by content type."""
+    apply_startup(ctx)
+    run_async(content_types_impl(ctx.obj.config, date_from=date_from, date_to=date_to))
+
+
+@analytics_app.command("hourly")
+def analytics_hourly(
+    ctx: typer.Context,
+    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
+    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
+) -> None:
+    """Hourly activity patterns."""
+    apply_startup(ctx)
+    run_async(hourly_impl(ctx.obj.config, date_from=date_from, date_to=date_to))
+
+
+@analytics_app.command("summary")
+def analytics_summary(ctx: typer.Context) -> None:
+    """Content generation summary."""
+    apply_startup(ctx)
+    run_async(summary_impl(ctx.obj.config))
+
+
+@analytics_app.command("daily")
+def analytics_daily(
+    ctx: typer.Context,
+    days: int = typer.Option(30, "--days", help="Number of days (default: 30)"),
+    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
+) -> None:
+    """Daily generation stats."""
+    apply_startup(ctx)
+    run_async(daily_impl(ctx.obj.config, days=days, pipeline_id=pipeline_id))
+
+
+@analytics_app.command("pipeline-stats")
+def analytics_pipeline_stats(
+    ctx: typer.Context,
+    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
+) -> None:
+    """Per-pipeline statistics."""
+    apply_startup(ctx)
+    run_async(pipeline_stats_impl(ctx.obj.config, pipeline_id=pipeline_id))
+
+
+@analytics_app.command("trending-topics")
+def analytics_trending_topics(
+    ctx: typer.Context,
+    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
+    limit: int = typer.Option(20, "--limit"),
+) -> None:
+    """Trending topics/keywords."""
+    apply_startup(ctx)
+    run_async(trending_topics_impl(ctx.obj.config, days=days, limit=limit))
+
+
+@analytics_app.command("trending-channels")
+def analytics_trending_channels(
+    ctx: typer.Context,
+    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
+    limit: int = typer.Option(20, "--limit"),
+) -> None:
+    """Top channels by activity."""
+    apply_startup(ctx)
+    run_async(trending_channels_impl(ctx.obj.config, days=days, limit=limit))
+
+
+@analytics_app.command("velocity")
+def analytics_velocity(
+    ctx: typer.Context,
+    days: int = typer.Option(30, "--days", help="Number of days (default: 30)"),
+) -> None:
+    """Message volume per day."""
+    apply_startup(ctx)
+    run_async(velocity_impl(ctx.obj.config, days=days))
+
+
+@analytics_app.command("peak-hours")
+def analytics_peak_hours(ctx: typer.Context) -> None:
+    """Peak activity hours."""
+    apply_startup(ctx)
+    run_async(peak_hours_impl(ctx.obj.config))
+
+
+@analytics_app.command("calendar")
+def analytics_calendar(
+    ctx: typer.Context,
+    limit: int = typer.Option(20, "--limit"),
+    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
+) -> None:
+    """Upcoming scheduled publications."""
+    apply_startup(ctx)
+    run_async(calendar_impl(ctx.obj.config, limit=limit, pipeline_id=pipeline_id))
+
+
+@analytics_app.command("trending-emojis")
+def analytics_trending_emojis(
+    ctx: typer.Context,
+    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
+    limit: int = typer.Option(20, "--limit"),
+) -> None:
+    """Trending emojis in messages."""
+    apply_startup(ctx)
+    run_async(trending_emojis_impl(ctx.obj.config, days=days, limit=limit))
+
+
+@analytics_app.command("channel", context_settings=_NEG_ID_POSITIONAL)
+def analytics_channel(
+    ctx: typer.Context,
+    channel_id: int = typer.Argument(..., help="Telegram channel_id (negative int)"),
+    days: int = typer.Option(30, "--days", help="Time window in days (default: 30)"),
+) -> None:
+    """Per-channel statistics overview."""
+    apply_startup(ctx)
+    run_async(channel_impl(ctx.obj.config, channel_id=channel_id, days=days))
+
+
+@analytics_app.command("channel-rating")
+def analytics_channel_rating(
+    ctx: typer.Context,
+    useful: AnalyticsUseful | None = typer.Option(None, "--useful", help="Filter by usefulness axis"),
+    genre: AnalyticsGenre | None = typer.Option(None, "--genre", help="Filter by genre axis"),
+    limit: int = typer.Option(50, "--limit", help="Max rows (default: 50)"),
+) -> None:
+    """Channel ratings (usefulness × genre)."""
+    apply_startup(ctx)
+    run_async(
+        channel_rating_impl(
+            ctx.obj.config,
+            useful=useful.value if useful else None,
+            genre=genre.value if genre else None,
+            limit=limit,
+        )
+    )
+
+
+@analytics_app.command("channel-rate", context_settings=_NEG_ID_POSITIONAL)
+def analytics_channel_rate(
+    ctx: typer.Context,
+    channel_id: int = typer.Argument(..., help="Telegram channel_id (bare positive int)"),
+    model: str | None = typer.Option(None, "--model", help="LLM model/provider name"),
+    sample_size: int = typer.Option(
+        40, "--sample-size", help="Number of recent posts to sample for the judge (default: 40)"
+    ),
+) -> None:
+    """Run the LLM judge on a channel and upsert its rating (usefulness × genre)."""
+    apply_startup(ctx)
+    run_async(
+        channel_rate_impl(
+            ctx.obj.config, channel_id=channel_id, model=model, sample_size=sample_size
+        )
+    )
