+++ src/cli/typer_commands.py
@@ -50,6 +50,8 @@ from src.cli.commands import settings as settings_cmd
 from src.cli.commands import test as test_cmd
 from src.cli.commands import translate as translate_cmd
 from src.cli.commands import worker as worker_cmd
+from src.cli.commands.analytics import analytics_app
+from src.cli.commands.channel import channel_app
 from src.cli.commands.collect import collect_app
 from src.cli.commands.common import (
     _NEG_ID_POSITIONAL,
@@ -68,6 +70,8 @@ from src.cli.commands.debug import debug_app
 from src.cli.commands.export import export_app
 from src.cli.commands.image import image_app
 from src.cli.commands.messages import messages_app
+from src.cli.commands.notification import notification_app
+from src.cli.commands.provider import provider_app
 from src.cli.commands.translate import translate_app
 from src.cli.typer_app import app, apply_startup, run_async
 from src.filters.criteria import DEFAULT_QUICK_SAMPLE_SIZE
@@ -285,512 +289,28 @@ app.add_typer(image_app, name="image")
 # provider → list / add / delete / probe / refresh / test-all
 # --------------------------------------------------------------------------- #
 
-provider_app = typer.Typer(no_args_is_help=True, help="LLM provider management")
 app.add_typer(provider_app, name="provider")
-
-
-@provider_app.command("list")
-def provider_list(ctx: typer.Context) -> None:
-    """List configured providers with models and status."""
-    apply_startup(ctx)
-    run_async(provider_cmd.list_impl(ctx.obj.config))
-
-
-@provider_app.command("add")
-def provider_add(
-    ctx: typer.Context,
-    name: str = typer.Argument(..., help="Provider name (e.g. openai, groq, anthropic)"),
-    api_key: str = typer.Option(..., "--api-key", help="API key"),
-    base_url: str | None = typer.Option(None, "--base-url", help="Custom base URL"),
-) -> None:
-    """Add or update a provider."""
-    apply_startup(ctx)
-    run_async(provider_cmd.add_impl(ctx.obj.config, name=name, api_key=api_key, base_url=base_url))
-
-
-@provider_app.command("delete")
-def provider_delete(
-    ctx: typer.Context,
-    name: str = typer.Argument(..., help="Provider name"),
-) -> None:
-    """Delete a provider."""
-    apply_startup(ctx)
-    run_async(provider_cmd.delete_impl(ctx.obj.config, name=name))
-
-
-@provider_app.command("probe")
-def provider_probe(
-    ctx: typer.Context,
-    name: str = typer.Argument(..., help="Provider name"),
-) -> None:
-    """Test provider connection."""
-    apply_startup(ctx)
-    run_async(provider_cmd.probe_impl(ctx.obj.config, name=name))
-
-
-@provider_app.command("refresh")
-def provider_refresh(
-    ctx: typer.Context,
-    name: str | None = typer.Argument(None, help="Provider name (default: all)"),
-) -> None:
-    """Refresh provider models."""
-    apply_startup(ctx)
-    run_async(provider_cmd.refresh_impl(ctx.obj.config, name=name))
-
-
-@provider_app.command("test-all")
-def provider_test_all(ctx: typer.Context) -> None:
-    """Test all configured providers."""
-    apply_startup(ctx)
-    run_async(provider_cmd.test_all_impl(ctx.obj.config))
 
 
 # --------------------------------------------------------------------------- #
 # notification → setup / status / delete / test / dry-run / set-account
 # --------------------------------------------------------------------------- #
 
-notification_app = typer.Typer(no_args_is_help=True, help="Personal notification bot management")
 app.add_typer(notification_app, name="notification")
-
-
-@notification_app.command("setup")
-def notification_setup(ctx: typer.Context) -> None:
-    """Create personal notification bot via BotFather."""
-    apply_startup(ctx)
-    run_async(notification_cmd.setup_impl(ctx.obj.config))
-
-
-@notification_app.command("status")
-def notification_status(ctx: typer.Context) -> None:
-    """Show notification bot status."""
-    apply_startup(ctx)
-    run_async(notification_cmd.status_impl(ctx.obj.config))
-
-
-@notification_app.command("delete")
-def notification_delete(ctx: typer.Context) -> None:
-    """Delete notification bot via BotFather."""
-    apply_startup(ctx)
-    run_async(notification_cmd.delete_impl(ctx.obj.config))
-
-
-@notification_app.command("test")
-def notification_test(
-    ctx: typer.Context,
-    message: str = typer.Option("Тестовое уведомление", "--message", help="Message text"),
-) -> None:
-    """Send a test notification message."""
-    apply_startup(ctx)
-    run_async(notification_cmd.test_impl(ctx.obj.config, message=message))
-
-
-@notification_app.command("dry-run")
-def notification_dry_run(ctx: typer.Context) -> None:
-    """Preview notification matches without sending."""
-    apply_startup(ctx)
-    run_async(notification_cmd.dry_run_impl(ctx.obj.config))
-
-
-@notification_app.command("set-account")
-def notification_set_account(
-    ctx: typer.Context,
-    phone: str = typer.Option(..., "--phone", help="Account phone number"),
-) -> None:
-    """Set account for notification bot."""
-    apply_startup(ctx)
-    run_async(notification_cmd.set_account_impl(ctx.obj.config, phone=phone))
 
 
 # --------------------------------------------------------------------------- #
 # analytics → top / content-types / hourly / summary / daily / pipeline-stats /
 #   trending-topics / trending-channels / velocity / peak-hours / calendar /
-#   trending-emojis / channel / channel-rating / channel-rate
-# --------------------------------------------------------------------------- #
 
-analytics_app = typer.Typer(no_args_is_help=True, help="Message analytics")
 app.add_typer(analytics_app, name="analytics")
-
-
-@analytics_app.command("top")
-def analytics_top(
-    ctx: typer.Context,
-    limit: int = typer.Option(20, "--limit", help="Number of results (default: 20)"),
-    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
-    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
-) -> None:
-    """Top messages by reactions."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.top_impl(ctx.obj.config, limit=limit, date_from=date_from, date_to=date_to))
-
-
-@analytics_app.command("content-types")
-def analytics_content_types(
-    ctx: typer.Context,
-    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
-    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
-) -> None:
-    """Engagement by content type."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.content_types_impl(ctx.obj.config, date_from=date_from, date_to=date_to))
-
-
-@analytics_app.command("hourly")
-def analytics_hourly(
-    ctx: typer.Context,
-    date_from: str | None = typer.Option(None, "--date-from", help="Start date (YYYY-MM-DD)"),
-    date_to: str | None = typer.Option(None, "--date-to", help="End date (YYYY-MM-DD)"),
-) -> None:
-    """Hourly activity patterns."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.hourly_impl(ctx.obj.config, date_from=date_from, date_to=date_to))
-
-
-@analytics_app.command("summary")
-def analytics_summary(ctx: typer.Context) -> None:
-    """Content generation summary."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.summary_impl(ctx.obj.config))
-
-
-@analytics_app.command("daily")
-def analytics_daily(
-    ctx: typer.Context,
-    days: int = typer.Option(30, "--days", help="Number of days (default: 30)"),
-    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
-) -> None:
-    """Daily generation stats."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.daily_impl(ctx.obj.config, days=days, pipeline_id=pipeline_id))
-
-
-@analytics_app.command("pipeline-stats")
-def analytics_pipeline_stats(
-    ctx: typer.Context,
-    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
-) -> None:
-    """Per-pipeline statistics."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.pipeline_stats_impl(ctx.obj.config, pipeline_id=pipeline_id))
-
-
-@analytics_app.command("trending-topics")
-def analytics_trending_topics(
-    ctx: typer.Context,
-    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
-    limit: int = typer.Option(20, "--limit"),
-) -> None:
-    """Trending topics/keywords."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.trending_topics_impl(ctx.obj.config, days=days, limit=limit))
-
-
-@analytics_app.command("trending-channels")
-def analytics_trending_channels(
-    ctx: typer.Context,
-    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
-    limit: int = typer.Option(20, "--limit"),
-) -> None:
-    """Top channels by activity."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.trending_channels_impl(ctx.obj.config, days=days, limit=limit))
-
-
-@analytics_app.command("velocity")
-def analytics_velocity(
-    ctx: typer.Context,
-    days: int = typer.Option(30, "--days", help="Number of days (default: 30)"),
-) -> None:
-    """Message volume per day."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.velocity_impl(ctx.obj.config, days=days))
-
-
-@analytics_app.command("peak-hours")
-def analytics_peak_hours(ctx: typer.Context) -> None:
-    """Peak activity hours."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.peak_hours_impl(ctx.obj.config))
-
-
-@analytics_app.command("calendar")
-def analytics_calendar(
-    ctx: typer.Context,
-    limit: int = typer.Option(20, "--limit"),
-    pipeline_id: int | None = typer.Option(None, "--pipeline-id"),
-) -> None:
-    """Upcoming scheduled publications."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.calendar_impl(ctx.obj.config, limit=limit, pipeline_id=pipeline_id))
-
-
-@analytics_app.command("trending-emojis")
-def analytics_trending_emojis(
-    ctx: typer.Context,
-    days: int = typer.Option(7, "--days", help="Number of days (default: 7)"),
-    limit: int = typer.Option(20, "--limit"),
-) -> None:
-    """Trending emojis in messages."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.trending_emojis_impl(ctx.obj.config, days=days, limit=limit))
-
-
-@analytics_app.command("channel", context_settings=_NEG_ID_POSITIONAL)
-def analytics_channel(
-    ctx: typer.Context,
-    channel_id: int = typer.Argument(..., help="Telegram channel_id (negative int)"),
-    days: int = typer.Option(30, "--days", help="Time window in days (default: 30)"),
-) -> None:
-    """Per-channel statistics overview."""
-    apply_startup(ctx)
-    run_async(analytics_cmd.channel_impl(ctx.obj.config, channel_id=channel_id, days=days))
-
-
-@analytics_app.command("channel-rating")
-def analytics_channel_rating(
-    ctx: typer.Context,
-    useful: AnalyticsUseful | None = typer.Option(None, "--useful", help="Filter by usefulness axis"),
-    genre: AnalyticsGenre | None = typer.Option(None, "--genre", help="Filter by genre axis"),
-    limit: int = typer.Option(50, "--limit", help="Max rows (default: 50)"),
-) -> None:
-    """Channel ratings (usefulness × genre)."""
-    apply_startup(ctx)
-    run_async(
-        analytics_cmd.channel_rating_impl(
-            ctx.obj.config,
-            useful=useful.value if useful else None,
-            genre=genre.value if genre else None,
-            limit=limit,
-        )
-    )
-
-
-@analytics_app.command("channel-rate", context_settings=_NEG_ID_POSITIONAL)
-def analytics_channel_rate(
-    ctx: typer.Context,
-    channel_id: int = typer.Argument(..., help="Telegram channel_id (bare positive int)"),
-    model: str | None = typer.Option(None, "--model", help="LLM model/provider name"),
-    sample_size: int = typer.Option(
-        40, "--sample-size", help="Number of recent posts to sample for the judge (default: 40)"
-    ),
-) -> None:
-    """Run the LLM judge on a channel and upsert its rating (usefulness × genre)."""
-    apply_startup(ctx)
-    run_async(
-        analytics_cmd.channel_rate_impl(
-            ctx.obj.config, channel_id=channel_id, model=model, sample_size=sample_size
-        )
-    )
 
 
 # --------------------------------------------------------------------------- #
 # channel → list / add / delete / toggle / collect / stats / refresh-types /
 #   refresh-meta / review-list / review-confirm / review-keep / import /
-#   add-bulk / list-for-import / tag (NESTED depth-2: list/add/delete/set/get)
-# --------------------------------------------------------------------------- #
 
-channel_app = typer.Typer(no_args_is_help=True, help="Channel management")
 app.add_typer(channel_app, name="channel")
-
-# Nested depth-2 group: ``channel tag`` is its own Typer added onto the channel
-# sub-app via ``add_typer`` — the exact path ``channel tag <action>`` is the
-# fragile frozen invariant of Wave 4.
-channel_tag_app = typer.Typer(no_args_is_help=True, help="Manage channel tags")
-channel_app.add_typer(channel_tag_app, name="tag")
-
-
-@channel_app.command("list")
-def channel_list(ctx: typer.Context) -> None:
-    """List channels."""
-    apply_startup(ctx)
-    run_async(channel_cmd.list_impl(ctx.obj.config))
-
-
-@channel_app.command("add", context_settings=_NEG_ID_POSITIONAL)
-def channel_add(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Username, link, or numeric ID"),
-) -> None:
-    """Add a channel."""
-    apply_startup(ctx)
-    run_async(channel_cmd.add_impl(ctx.obj.config, identifier=identifier))
-
-
-@channel_app.command("delete", context_settings=_NEG_ID_POSITIONAL)
-def channel_delete(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Channel pk, channel_id, or @username"),
-) -> None:
-    """Delete a channel."""
-    apply_startup(ctx)
-    run_async(channel_cmd.delete_impl(ctx.obj.config, identifier=identifier))
-
-
-@channel_app.command("toggle", context_settings=_NEG_ID_POSITIONAL)
-def channel_toggle(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Channel pk, channel_id, or @username"),
-) -> None:
-    """Toggle channel active state."""
-    apply_startup(ctx)
-    run_async(channel_cmd.toggle_impl(ctx.obj.config, identifier=identifier))
-
-
-@channel_app.command("collect", context_settings=_NEG_ID_POSITIONAL)
-def channel_collect(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Channel pk, channel_id, or @username"),
-    full: bool = typer.Option(False, "--full", help="Explicitly backfill the full channel history"),
-) -> None:
-    """Collect a single channel one-shot."""
-    apply_startup(ctx)
-    run_async(channel_cmd.collect_impl(ctx.obj.config, identifier=identifier, full=full))
-
-
-@channel_app.command("stats", context_settings=_NEG_ID_POSITIONAL)
-def channel_stats(
-    ctx: typer.Context,
-    identifier: str | None = typer.Argument(None, help="Channel pk, channel_id, or @username"),
-    all_channels: bool = typer.Option(False, "--all", help="Collect stats for all active channels"),
-    max_channels: int | None = typer.Option(
-        None, "--max-channels", help="Maximum active channels to process in this bounded stats-all run"
-    ),
-) -> None:
-    """Collect channel stats."""
-    apply_startup(ctx)
-    run_async(
-        channel_cmd.stats_impl(
-            ctx.obj.config, all_channels=all_channels, identifier=identifier, max_channels=max_channels
-        )
-    )
-
-
-@channel_app.command("refresh-types")
-def channel_refresh_types(ctx: typer.Context) -> None:
-    """Re-resolve channel types for all active channels."""
-    apply_startup(ctx)
-    run_async(channel_cmd.refresh_types_impl(ctx.obj.config))
-
-
-@channel_app.command("refresh-meta", context_settings=_NEG_ID_POSITIONAL)
-def channel_refresh_meta(
-    ctx: typer.Context,
-    identifier: str | None = typer.Argument(None, help="Channel pk, channel_id, or @username (omit for all)"),
-    all_channels: bool = typer.Option(False, "--all", help="Refresh metadata for all active channels"),
-) -> None:
-    """Refresh channel metadata."""
-    apply_startup(ctx)
-    run_async(
-        channel_cmd.refresh_meta_impl(ctx.obj.config, all_channels=all_channels, identifier=identifier)
-    )
-
-
-@channel_app.command("review-list")
-def channel_review_list(ctx: typer.Context) -> None:
-    """List channels quarantined for review."""
-    apply_startup(ctx)
-    run_async(channel_cmd.review_list_impl(ctx.obj.config))
-
-
-@channel_app.command("review-confirm", context_settings=_NEG_ID_POSITIONAL)
-def channel_review_confirm(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Channel pk, channel_id, or @username"),
-) -> None:
-    """Confirm a quarantined channel is dead and deactivate it."""
-    apply_startup(ctx)
-    run_async(channel_cmd.review_confirm_impl(ctx.obj.config, identifier=identifier))
-
-
-@channel_app.command("review-keep", context_settings=_NEG_ID_POSITIONAL)
-def channel_review_keep(
-    ctx: typer.Context,
-    identifier: str = typer.Argument(..., help="Channel pk, channel_id, or @username"),
-) -> None:
-    """Clear a channel's review flag and keep it active."""
-    apply_startup(ctx)
-    run_async(channel_cmd.review_keep_impl(ctx.obj.config, identifier=identifier))
-
-
-@channel_app.command("import", context_settings=_NEG_ID_POSITIONAL)
-def channel_import(
-    ctx: typer.Context,
-    source: str = typer.Argument(..., help="Path to .txt/.csv file, or comma-separated identifiers"),
-) -> None:
-    """Bulk-import channels from a file or identifier list."""
-    apply_startup(ctx)
-    run_async(channel_cmd.import_impl(ctx.obj.config, source=source))
-
-
-@channel_app.command("add-bulk")
-def channel_add_bulk(
-    ctx: typer.Context,
-    phone: str = typer.Option(..., "--phone", help="Account phone"),
-    dialog_ids: str = typer.Option(..., "--dialog-ids", help="Comma-separated dialog IDs to add as channels"),
-) -> None:
-    """Add channels from an account's dialogs by id list."""
-    apply_startup(ctx)
-    run_async(channel_cmd.add_bulk_impl(ctx.obj.config, phone=phone, dialog_ids=dialog_ids))
-
-
-@channel_app.command("list-for-import")
-def channel_list_for_import(
-    ctx: typer.Context,
-    as_json: bool = typer.Option(False, "--json", help="Output as JSON instead of a table"),
-) -> None:
-    """List dialogs with an already-added flag."""
-    apply_startup(ctx)
-    run_async(channel_cmd.list_for_import_impl(ctx.obj.config, as_json=as_json))
-
-
-# ---- nested: channel tag <action> ---------------------------------------- #
-
-
-@channel_tag_app.command("list")
-def channel_tag_list(ctx: typer.Context) -> None:
-    """List all channel tags."""
-    apply_startup(ctx)
-    run_async(channel_cmd._tag_impl(ctx.obj.config, "list"))
-
-
-@channel_tag_app.command("add")
-def channel_tag_add(
-    ctx: typer.Context,
-    name: str = typer.Argument(..., help="Tag name"),
-) -> None:
-    """Create a channel tag."""
-    apply_startup(ctx)
-    run_async(channel_cmd._tag_impl(ctx.obj.config, "add", name=name))
-
-
-@channel_tag_app.command("delete")
-def channel_tag_delete(
-    ctx: typer.Context,
-    name: str = typer.Argument(..., help="Tag name"),
-) -> None:
-    """Delete a channel tag."""
-    apply_startup(ctx)
-    run_async(channel_cmd._tag_impl(ctx.obj.config, "delete", name=name))
-
-
-@channel_tag_app.command("set")
-def channel_tag_set(
-    ctx: typer.Context,
-    pk: int = typer.Argument(..., help="Channel primary key"),
-    tags: str = typer.Argument(..., help="Comma-separated tag names"),
-) -> None:
-    """Replace a channel's tags."""
-    apply_startup(ctx)
-    run_async(channel_cmd._tag_impl(ctx.obj.config, "set", pk=pk, tags=tags))
-
-
-@channel_tag_app.command("get")
-def channel_tag_get(
-    ctx: typer.Context,
-    pk: int = typer.Argument(..., help="Channel primary key"),
-) -> None:
-    """Show a channel's tags."""
-    apply_startup(ctx)
-    run_async(channel_cmd._tag_impl(ctx.obj.config, "get", pk=pk))
 
 
 # --------------------------------------------------------------------------- #
